+++ CHANGELOG.rst
@@ -33,6 +33,7 @@ dev
    - Allow plotting densities in viewer functions for all unit cell types
    - Add an option to plot both spin channels in band structure plots
    - Add DOS calculation and plot functions
+   - Add an isovalue keyword to the density viewer
    - Add pass-through keyword arguments in the Cell creation
    - Add view and write class methods to Atoms, SCF, and KPoints objects
    - Set default values for uninitialized SCF attributes to None